23/10/2023 15:57

+++ Meet-My-Lecture/src/main/java/com/springboot/meetMyLecturer/entity/Notification.java
@@ -16,7 +16,7 @@ public class Notification {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "noti_id")
-    private int notiId;
+    private Long notiId;
 
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false)
